Replace decodeObject with json object direct lookup
Using `decodeObject` on the JSON in the `<|` operator is causing a heavy
performance hit. This operator is the base operator that all the other
ones rely on. That means that this one is called every time any other
one it called. The `decodeObject` function is looping through every
key-value pair in the object and calling `decode` on the value. Then
after this function returns we are using the dictionary subscript to
pull only one object from that dictionary, meaning we just decoded a
bunch of values that we didn't need to decode. This unused decoding is
happening so often that we're seeing performance hits of 100 times or
more.

Pulling out the interal JSON from the given JSON first allows us to
decode only what we're looking for each time.